Time: 0 ms (100%), Space: 7.9 MB (19.33%) - LeetHub

+++ 1406-subtract-the-product-and-sum-of-digits-of-an-integer/1406-subtract-the-product-and-sum-of-digits-of-an-integer.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     int subtractProductAndSum(int n) {
         size_t product = 1;
-        unsigned short sum = 0;
+        unsigned int sum = 0;
 
         while (n > 0)
         {
